error handling

+++ routes/ads.js
@@ -12,10 +12,10 @@ var multipartMiddleware = multiPart();
 /*jslint sloppy: true*/
 /*jslint nomen: true*/
 router.use(methodOverride(function (req) {
+
     if (req.body && typeof req.body === 'object' && '_method' in req.body) {
-        var method = req.body._method;
         delete req.body._method;
-        return method;
+        return req.body._method;
     }
 }));
 
@@ -28,32 +28,34 @@ function photoHandler(uploadedImage) {
     if (uploadedImage.type == 'image/png' || uploadedImage.type == 'image/jpeg') {
 
         var imageName = Math.random() + '.jpg';
-        console.log(imageName);
+
         im.resize({
+
             srcPath: uploadedImage.path,
             dstPath: './public/images/small/' + imageName,
             width: 200
+
         }, function (err) {
-            if (err) {
-                throw err;
-            }
+            if (err) { return err; }
         });
 
         im.resize({
+
             srcPath: uploadedImage.path,
             dstPath: './public/images/big/' + imageName,
             width: 600
+
         }, function (err) {
-            if (err) {
-                throw err;
-            }
+            if (err) { throw err; }
         });
+
         return imageName;
     }
     return false;
 }
 
 function checkAuth(req, res, next) {
+
     if (!req.session.user_id) {
         req.flash('info', 'Please log in');
         res.redirect('/users/login');
@@ -63,9 +65,11 @@ function checkAuth(req, res, next) {
 }
 
 function imageCaptcha(captcha) {
+
     var img,
         imgbase64,
         p = new Captchapng(80, 30, captcha); // width,height,numeric captcha
+
     p.color(0, 0, 0, 0);  // First color: background (red, green, blue, alpha)
     p.color(80, 80, 80, 255); // Second color: paint (red, green, blue, alpha)
     img = p.getBase64();
@@ -73,10 +77,11 @@ function imageCaptcha(captcha) {
     return imgbase64;
 }
 
-router.get('/newad', checkAuth, function (req, res) {
+router.get('/newad', checkAuth, function (req, res, next) {
 
     req.db.get('users').findById(req.session.user_id, function (err, doc) {
-        if (err || !doc) {
+        if (err) { return next(err); }
+        if (!doc) {
             res.redirect('/users/profile');
             return;
         }
@@ -97,7 +102,9 @@ router.get('/newad', checkAuth, function (req, res) {
 router.param('id', function (req, res, next, id) {
 
     req.db.get('ads').findById(id, function (err, ad) {
-        if (err || !ad) {
+        if (err) { return next(err); }
+
+        if (!ad) {
             req.flash('info', req.app.locals.i18n('noAd'));
             res.redirect('/users/profile');
         } else {
@@ -108,14 +115,22 @@ router.param('id', function (req, res, next, id) {
 });
 
 
-router.get('/:id', function (req, res) {
+router.get('/:id', function (req, res, next) {
+
     var db = req.db,
         userCol = db.get('users'),
         adCol = db.get('ads'),
         categoryCol = db.get('categories');
+
     adCol.findById(req.id, function (err, ad) {
+        if (err) { return next(err); }
+
         categoryCol.findById(ad.category_id, function (err, category) {
+            if (err) { return next(err); }
+
             userCol.findById(ad.user_id, function (err, user) {
+                if (err) { return next(err); }
+
                 res.render('ad/show', {
                     curPage: '/ads/' + req.id,
                     ad : ad,
@@ -129,16 +144,24 @@ router.get('/:id', function (req, res) {
 });
 
 
-router.get('/:id/edit', checkAuth, function (req, res) {
+router.get('/:id/edit', checkAuth, function (req, res, next) {
+
     var db = req.db,
         adCol = db.get('ads'),
         userCol = db.get('users'),
         categoryCol = db.get('categories'),
         captcha = parseInt(Math.random() * 9000 + 1000, 10);
+
     adCol.findById(req.id, function (err, ad) {
+        if (err) { return next(err); }
+
         categoryCol.findById(ad.category_id, function (err, category) {
+            if (err) { return next(err); }
+
             userCol.findById(ad.user_id, function (err, user) {
+                if (err) { return next(err); }
                 req.session.captcha = captcha;
+
                 res.render('ad/newad', {
                     curPage: '/ads/' + req.id + '/edit',
                     user : user,
@@ -154,7 +177,7 @@ router.get('/:id/edit', checkAuth, function (req, res) {
 });
 
 
-var adCallback = function (req, res) {
+var adCallback = function (req, res, next) {
 
     if (req.body.captcha != req.session.captcha) {
         req.flash('info', req.app.locals.i18n('noCaptcha'));
@@ -181,6 +204,7 @@ var adCallback = function (req, res) {
 
         if (req.files[fieldName].name != '') {
             imageName = photoHandler(req.files[fieldName]);
+
             if (imageName === false) {
                 req.flash('info', req.app.locals.i18n('incorrectImage'));
                 res.redirect(req.id !== undefined ? '/ads/' + req.id + '/edit' : '/ads/newad');
@@ -195,9 +219,11 @@ var adCallback = function (req, res) {
 
         // removing old pictures
         adCol.findOne({ _id: req.id, user_id: new ObjectId(req.session.user_id) }, function (err, doc) {
-            if (err || !doc) { return; }
+            if (err || !doc) { return next(err); }
+
             for (i = 1; i < 3; i++) {
                 fieldName = 'image' + i;
+
                 if (req.files[fieldName].name != '' && doc[fieldName] != undefined) {
                     fs.unlink('./public/images/big/' + doc[fieldName]);
                     fs.unlink('./public/images/small/' + doc[fieldName]);
@@ -211,9 +237,8 @@ var adCallback = function (req, res) {
     } else {
 
         db.get('categories').findById(req.body.category, function (err, doc) {
-            if (err || !doc) {
-                res.redirect('/ads/newad');
-            }
+            if (err || !doc) { return next(err); }
+
             if (doc) {
                 colObject.category_id = new ObjectId(req.body.category);
                 adCol.insert(colObject);
@@ -228,7 +253,7 @@ router.post('/addad', checkAuth, multipartMiddleware, adCallback);
 
 router.post('/:id/adedit', checkAuth, multipartMiddleware, adCallback);
 
-router.get('/:id/imgdel', checkAuth, function (req, res) {
+router.get('/:id/imgdel', checkAuth, function (req, res, next) {
 
     var i,
         imageNumber,
@@ -237,10 +262,11 @@ router.get('/:id/imgdel', checkAuth, function (req, res) {
         imgObject = {};
 
     adCol.findOne({ _id: req.id, user_id: new ObjectId(req.session.user_id) }, function (err, doc) {
-        if (err) { throw err; }
-        if (!doc) { return; }
+        if (err || !doc) { return next(err); }
+
         for (i = 1; i < 3; i++) {
             imageNumber = 'image' + i;
+
             if (req.query.img === doc[imageNumber]) {
                 imgObject[imageNumber] = 1;
                 fs.unlink('./public/images/big/' + req.query.img);
@@ -253,7 +279,7 @@ router.get('/:id/imgdel', checkAuth, function (req, res) {
 });
 
 
-router.delete('/:id', checkAuth, function (req, res) {
+router.delete('/:id', checkAuth, function (req, res, next) {
 
     var i,
         fieldName,
@@ -261,13 +287,15 @@ router.delete('/:id', checkAuth, function (req, res) {
         adCol = db.get('ads');
 
     adCol.findOne({ _id: req.id, user_id: new ObjectId(req.session.user_id) }, function (err, doc) {
-        if (err) { return err; }
-        if (!doc) { return; }
+        if (err || !doc) { return next(err); }
+
         adCol.removeById(doc._id, function (err) {
-            if (err) { throw err; }
+            if (err) { return next(err); }
         });
+
         for (i = 1; i < 3; i++) {
             fieldName = 'image' + i;
+
             if (doc[fieldName] != undefined) {
                 fs.unlink('./public/images/big/' + doc[fieldName]);
                 fs.unlink('./public/images/small/' + doc[fieldName]);
